Uploading created script
Script I wrote to verify the integrity of Sisense data security rules. This generates a list of security rules. If a security rule has been newly added, the rule is recorded in the security baseline. If a rule is modified, an email will be sent stating the rule that has been modified. Useful for auditing sensitive data policies.