Create StateParkScraper
Accepts 'states.txt' with a list of states on each individual line. Returns a CSV file of all documented state and federal parks within each state, along with an address lookup performed by the Google Places API for each park. 

I used this to create a Google mymaps rendering of all the parks when planning a road trip.